Add all that is needed to handle game states and scene loading
This commit introduces a Scene Manager which contains logic
to load and unload scenes based on Levels. It is quite limited:
the currently loaded scene is always automatically unloaded before
a new one is loaded.

It introduces also a basic life system with a Game Over mechanism
that allows to restart the game.

It is all very simple and dirty due to lack of time and experience in scene managment and async asset loading. But hey, it will do for now.